fix : 버그 수정을 위한 코드 공유

+++ app/src/main/java/com/example/diadi/view/SearchActivity.kt
@@ -48,7 +48,7 @@ class SearchActivity : AppCompatActivity() {
         binding.rvList.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
         binding.rvList.adapter = searchAdapter
 
-        // 1-2 사용자가 키워드 검색 후 장소 아이템 클릭하면 발생하는 이벤트!
+        // 사용자가 키워드 검색 후 장소 아이템 클릭하면 발생하는 이벤트!
         searchAdapter.setItemClickListener(object: SearchAdapter.OnItemClickListener {
             override fun onClick(v: View, position: Int) {
                 val mapPoint = MapPoint.mapPointWithGeoCoord(searchItems[position].y, searchItems[position].x)
@@ -63,12 +63,6 @@ class SearchActivity : AppCompatActivity() {
                     intent.putExtra("category", searchItems[position].category)
                     intent.putExtra("address", searchItems[position].address)
                 }
-
-                // 1-3 여기서 이 이벤트가 한번 발생하고 나서(선택한 장소가 있을 때), 저장 버튼을 누르면 그 장소 정보를 다시 addActivity로 반환
-                // 이러면 저장 버튼이 눌렸을때 event listener가 하나 있어야 하고.
-                // 그 리스너에서는 searchItems[position]을 반환하면 돼..! 그냥 return searchItems[position]
-
-                // 아이템 클릭 시 저장 버튼이 활성화되고, 저장 버튼을 누르면 putExtra로 위도와 경도 데이터를 내보내게 설계했음 (완료)
             }
         })
 
@@ -135,8 +129,8 @@ class SearchActivity : AppCompatActivity() {
                     document.place_name,
                     document.category_group_name,
                     document.road_address_name,
-                    document.x.toDouble(),
-                    document.y.toDouble()
+                    document.x,
+                    document.y
                 )
 
                 searchItems.add(item)
@@ -158,7 +152,7 @@ class SearchActivity : AppCompatActivity() {
         val point = MapPOIItem()
         point.apply {
             itemName = document.place_name
-            mapPoint = MapPoint.mapPointWithGeoCoord(document.y.toDouble(), document.x.toDouble())
+            mapPoint = MapPoint.mapPointWithGeoCoord(document.y, document.x)
             markerType = MapPOIItem.MarkerType.RedPin
             selectedMarkerType = MapPOIItem.MarkerType.BluePin
         }
